More AV sync-up in runtime
os/linux/driver/opencpi.c: warning
runtime/application/include/OcpiApplicationApi.h: merged dumpProperties
runtime/application/src/OcpiApplication.cxx: ditto
runtime/application/include/OcpiApplication.h: add const
runtime/container/include/ContainerLauncher.h: move Singleton to OU::
runtime/container/include/ContainerWorker.h: sanitize comments
runtime/container/include/OcpiContainerApi.h: expose baseType()
runtime/container/src/ContainerProperty.cxx: ditto
runtime/container/src/ContainerApplication.cxx: sanitize comments
runtime/container/src/ContainerPort.cxx: sanitize comments
runtime/ctests/include/test_utilities.h: better type, additional member inits
runtime/ctests/src/RTxTest_main.cxx: ditto
runtime/ctests/src/run_tests.sh: support RPM mode too
runtime/ctests/src/test_optional_ports_main.cxx: improve message
runtime/ctests/src/test_utilities.cxx: preincrement interators
runtime/ctests/src/test_zero_copy_io_main.cxx: valgrind hook, replace (void) with /*var*/
runtime/dataplane/msg_driver_interface/include/DtMsgDriver.h: move singleton to OU::
runtime/dataplane/msg_drivers/Makefile.ocpi.imports: use OCPI_TARGET_OS not OCPI_OS
runtime/dataplane/msg_drivers/corba/src/OcpiCORBATransport.cxx: sanitize, mem leak, preincrement
runtime/dataplane/msg_drivers/dds/src/OcpiDDSTransport.cxx: insert comment about cppcheck catch
runtime/dataplane/rdma_driver_interface/interfaces/include/DtDriver.h: move Singleton to OU::
runtime/hdl-support/src/HdlSimServer.cxx: warnings
runtime/library/include/OcpiLibraryManager.h: sanitize
runtime/library/src/LibrarySimple.cxx: move Singleton to OU::
runtime/library/src/OcpiLibraryManager.cxx: be more careful about NULL chars and bufsize
runtime/rcc/include/RCC_Worker.h: sanitize, add getLength for symmetry, shadowing
runtime/rcc/src/RccWorker.cxx: error checks, shadowing, debug, indent
runtime/remote/src/RemoteDriver.cxx: more asserts
runtime/util/Makefile.ocpi.imports: use OCPI_TARGET_OS, not OCPI_OS
runtime/util/bin/src/fasttimed_main.cxx: add error checks
runtime/util/bin/src/timeCvt_main.cxx: init warning
runtime/util/driver/include/OcpiDriverManager.h: move Singleton, to OU::
runtime/util/driver/src/OcpiDriverManager.cxx: ditto
runtime/util/ezxml/src/OcpiUtilEzxml.cxx: sanitize
runtime/util/logger/include/OcpiLoggerLogger.h: sanitize
runtime/util/misc/include/OcpiUtilMisc.h: add string comparitors, add Singleton
runtime/util/misc/src/OcpiUtilException.cxx: add missing va_end
runtime/util/misc/src/OcpiUtilMisc.cxx: add error checks
runtime/util/property/src/OcpiUtilDataTypes.cxx: warnings, error checks
runtime/util/property/src/OcpiUtilImplementation.cxx: sanitize, leak
runtime/util/property/src/OcpiUtilProtocol.cxx: error checks, fix typo in msg
runtime/util/property/src/OcpiUtilValue.cxx: error check
runtime/util/time/include/cycle.h: sanitize
runtime/util/timeEmit/include/OcpiTimeEmitOutputFormat.h: missing init
tests/test-os/src/test-timer.cxx: add more timer slack for VMs etc.
runtime/util/fasttime/src/calibration.cxx: use inttypes
runtime/util/http/src/OcpiUtilHttpFs.cxx: sanitize

+++ runtime/hdl-support/src/HdlSimServer.cxx
@@ -500,7 +500,7 @@ namespace OCPI {
 	      int en = errno;
 	      x += m_dir;
 	      x += "\n";
-	      ocpiCheck(write(2, x.c_str(), x.length()) == x.length());
+	      ocpiCheck(write(2, x.c_str(), x.length()) == (ssize_t)x.length());
 	      _exit(10 + en);
 	    }
 	    {
@@ -508,7 +508,7 @@ namespace OCPI {
 	      if (fd < 0) {
 		std::string x("Error: Cannot create sim.out file for simulation output.\n");
 		int en = errno;
-		ocpiCheck(write(2, x.c_str(), x.length()) == x.length());
+		ocpiCheck(write(2, x.c_str(), x.length()) == (ssize_t)x.length());
 		_exit(10 + en);
 	      }
 	      if (dup2(fd, 1) < 0 ||
